Chore: ADAPT-3747 Rename structureIcons.tsx → StructureIcons.tsx
Match the PascalCase convention of the other components/course files;
update imports in CourseStructureTree and CourseStructureMapView.
