+++ new-ui-source/src/components/course/StructureIcons.tsx
